pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "creative-automation-cli"
version = "0.1.0"
description = "POC CLI for creative automation campaign generation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.8.0",
  "PyYAML>=6.0.1",
  "Pillow>=10.4.0",
  "jsonschema>=4.23.0",
  "requests>=2.32.0",
]

[project.optional-dependencies]
gemini = [
  "google-genai>=1.27.0",
]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.4.0",
]

[project.scripts]
generate_campaign = "creative_automation_cli.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]  # line-length enforced by formatter, not linter

[tool.ruff.lint.isort]
known-first-party = ["creative_automation_cli"]
